Move borrowed_box to its own module
